requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
# Optional: prebuilt FAISS wheels on Windows (if default fails)
# --extra-index-url https://github.com/kyamagu/faiss-wheels/releases

# Core scientific stack (Python 3.12 compatible)
numpy==1.26.4
pandas==2.1.4
scipy==1.13.1
scikit-learn==1.4.2

# Deep learning
torch==2.2.2
sentence-transformers==2.6.1
# align with sentence-transformers 2.6.1 requirement (>=4.41.0)
transformers==4.41.0

# Vector search
faiss-cpu==1.7.4

# Sparse retrieval
rank-bm25==0.2.2
splade==2.1.0

# Database / vector store
supabase==1.0.3
vecs==0.4.0

# API + utilities
openai==1.3.0
python-dotenv==1.0.0
python-dateutil==2.9.0.post0
tqdm==4.66.4
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
